adjust download file maker to allow isWebServiceFile option
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/ApiCommonWorkflow/trunk@45392 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ Main/lib/perl/WorkflowSteps/DownloadFileMaker.pm
@@ -43,35 +43,42 @@ sub run {
   my $relativeDir = $self->getParamValue('relativeDir');
   my $fileType = $self->getParamValue('fileType');
   my $dataName = $self->getParamValue('dataName');
-  my $descripString= $self->getParamValue('descripString');
+  my $descripString = $self->getParamValue('descripString');
+  my $isWebServiceFile = $self->getBooleanParamValue('isWebServiceFile');
 
   my $websiteFilesDir = $self->getWebsiteFilesDir($test);
   
-  my $organismNameForFiles = $self->getOrganismInfo($test, $organismAbbrev)->getNameForFiles();
-  my $speciesNameForFiles = $self->getOrganismInfo($test, $organismAbbrev)->getSpeciesNameForFiles();
-  my $nameForFiles = $self->getIsSpeciesLevel()?  $speciesNameForFiles:  $organismNameForFiles;
-  my $outputDir = "$websiteFilesDir/$relativeDir/$nameForFiles/$fileType";
-  
-  $dataName = "_$dataName" if $dataName; # gff does not use $dataName, so allow it to be empty
+  my ($outputDir, $downloadFile, $descripFile, $descripFileCmd);
 
-  my $downloadFile = "$outputDir/$projectName-${projectVersion}_${nameForFiles}_$dataName.$fileType";
-  my $descripFile = "$outputDir/.$projectName-${projectVersion}_${nameForFiles}_$dataName.$fileType.desc";
-  my $descripFileCmd =  "writeDownloadFileDecripWithDescripString --descripString '$descripString' --outputFile $descripFile";
+  if ($isWebServiceFile) {
+      $downloadFile = "$websiteFilesDir/$relativeDir/$organismAbbrev/$fileType/$dataName.$fileType";
+  } else {
+      my $organismNameForFiles = $self->getOrganismInfo($test, $organismAbbrev)->getNameForFiles();
+      my $speciesNameForFiles = $self->getOrganismInfo($test, $organismAbbrev)->getSpeciesNameForFiles();
+      my $nameForFiles = $self->getIsSpeciesLevel()?  $speciesNameForFiles:  $organismNameForFiles;
+      my $outputDir = "$websiteFilesDir/$relativeDir/$nameForFiles/$fileType";
+      
+      $dataName = "_$dataName" if $dataName; # gff does not use $dataName, so allow it to be empty
+
+      $downloadFile = "$outputDir/$projectName-${projectVersion}_${nameForFiles}_$dataName.$fileType";
+      $descripFile = "$outputDir/.$projectName-${projectVersion}_${nameForFiles}_$dataName.$fileType.desc";
+      $descripFileCmd =  "writeDownloadFileDecripWithDescripString --descripString '$descripString' --outputFile $descripFile";
+  }
 
   my $downloadFileCmd =  $self->getDownloadFileCmd($downloadFile, $test);
 
   if($undo){
     $self->runCmd(0, "rm -f $downloadFile") unless $downloadFileCmd eq 'NONE';
-    $self->runCmd(0, "rm -f $descripFile");
+    $self->runCmd(0, "rm -f $descripFile") unless $isWebServiceFile;
   }else {
       $self->error("Output file '$downloadFile' already exists") if -e $downloadFile;
-      $self->error("Output file '$descripFile' already exists") if -e $descripFile;
+      $self->error("Output file '$descripFile' already exists") if !$isWebServiceFile && -e $descripFile;
       if ($test) {
 	  $self->runCmd(0, "echo test > $downloadFile") unless $downloadFileCmd eq 'NONE';
-	  $self->runCmd(0, "echo test > $descripFile")
+	  $self->runCmd(0, "echo test > $descripFile")  unless $isWebServiceFile;
       }else {
 	  $self->runCmd($test, $downloadFileCmd) unless $downloadFileCmd eq 'NONE';
-	  $self->runCmd($test, $descripFileCmd);
+	  $self->runCmd($test, $descripFileCmd)  unless $isWebServiceFile;
       }
   }
 }
